+++ src/js/opalbox.jquery.js
@@ -4,7 +4,7 @@
  * javascript.
  *
  * @author Angel M (@laux_es)
- * @version 0.0.2rc
+ * @version 1.0.0
  * @license MIT License
  */
 ;(function ($, window, document, undefined) {
@@ -15,6 +15,7 @@
     // Store base
     this.element = element;
     this._name = pluginName;
+    this.requirements = false;
     this._defaults = $.fn.opalBox.defaults;
 
     // Get options
@@ -28,6 +29,7 @@
   $.extend(OpalBox.prototype, {
     // Initialization
     init: function () {
+      this._initOpal();
       this._createBox();
       this._buildCache();
       this._bindEvents();
@@ -36,6 +38,16 @@
     destroy: function() {
       this._unbindEvents();
       this.$element.removeData();
+    },
+    // Load compile OpalBox library
+    _initOpal: function() {
+      if (typeof Opal !== 'object' ||
+          typeof Opal.modules["opal-parser"] !== 'function'){
+        this._notLoadedError();
+      } else {
+        this.requirements = true;
+        Opal.load('opal-parser');
+      }
     },
     // Create the HTML elements
     _createBox: function() {
@@ -96,7 +108,10 @@
         self.$code.trigger('input');
       }
     },
-
+    // Show an error because Opal libraries are not loaded.
+    _notLoadedError: function() {
+      console.warn("Required Opal libraries are not loaded");
+    },
     // Unbind events that trigger methods
     unbindEvents: function() {
       // Unbind click execute event.
@@ -108,6 +123,12 @@
     },
     // Execute the code and show the output
     execute: function() {
+      if (!this.requirements) {
+        this._notLoadedError();
+        this.$result.html("<p class='empty'>Libraries are not loaded.<p>");
+        return;
+      }
+
       // Show button as "Running"
       this._setButtonLoading();
 
@@ -117,7 +138,7 @@
           res;
 
       // Show a loading text
-      this.$result.html("<p>...</p>");
+      this.$result.html("");
 
       // Modify the log to capture "puts" sentences
       try {
@@ -133,8 +154,6 @@
       // Return to normal console
       window.console.log = this.standardLog;
 
-      // Clean output
-      this.$result.html("");
       // Show result
       if (error === undefined){
         this._showResult(res);